+++ beTheStreet/src/stdio/kiteDream/module/feedback/service/FeedbackService.java
@@ -12,6 +12,6 @@ public interface FeedbackService {
 
 	public boolean deleteFeedback(String id);
 
-	public long getCount();
+	public int getCount();
 
 }
